# src/main/java/com/chubb/policyoverview/api/dto/request/FlagPoliciesRequestDto.java
package com.chubb.policyoverview.api.dto.request;

import jakarta.validation.constraints.NotEmpty;
import jakarta.validation.constraints.NotNull;
import java.util.List;
import java.util.UUID;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class FlagPoliciesRequestDto {

    @NotEmpty
    private List<@NotNull UUID> policyIds;
}